Separate data from main service
